Criando helpers dentro do Component alterar cronograma
