+++ src/components/screens/helper.js
@@ -17,7 +17,6 @@ import { truncarString } from "../../helpers/utilities";
 import { TIPO_PERFIL, TIPO_SOLICITACAO } from "constants/shared";
 import { usuarioEhEscola } from "../../helpers/utilities";
 import { STATUS_ALIMENTO } from "./const";
-import moment from "moment";
 
 export const ALT_CARDAPIO = "ALT_CARDAPIO";
 export const ALT_CARDAPIO_CEMEI = "ALT_CARDAPIO_CEMEI";
@@ -253,45 +252,4 @@ export const getDataHomologacao = logs => {
     logs
   );
   return arr[0] ? arr[0].criado_em : "--";
-};
-
-export const prepararPayloadCronograma = (cronograma, values) => {
-  let etapas = prepararPayloadEtapas(cronograma, values);
-  return {
-    cronograma: cronograma.uuid,
-    motivo: values.motivos,
-    etapas: etapas,
-    justificativa: values.justificativa
-  };
-};
-
-export const prepararPayloadEtapas = (cronograma, values) => {
-  let etapas = [];
-  if (!values.motivos) {
-    return;
-  }
-  if (!values.motivos.includes("outros")) {
-    etapas = cronograma.etapas.map(etapa => {
-      let etapas_payload = {
-        uuid: etapa.uuid
-      };
-      if (
-        values.motivos.includes("ALTERAR_DATA_ENTREGA") &&
-        values[`data_programada_${etapa.uuid}`]
-      ) {
-        etapas_payload["nova_data_programada"] = moment(
-          values[`data_programada_${etapa.uuid}`]
-        ).format("YYYY-MM-DD");
-      }
-      if (
-        values.motivos.includes("ALTERAR_QTD_ALIMENTO") &&
-        values[`quantidade_total_${etapa.uuid}`]
-      ) {
-        etapas_payload["nova_quantidade"] =
-          values[`quantidade_total_${etapa.uuid}`];
-      }
-      return etapas_payload;
-    });
-    return etapas;
-  }
 };
